+++ src/fncs/pages/index_files/libs.ignore/pageFilter.js
@@ -1,0 +1,6 @@
+/**
+ * pageFilter.js
+ */
+module.exports = function(app, px, pj){
+	var it79 = require('iterate79');
+}
